Rename queue subproject to tasks
