--- services/queue/build.gradle
@@ -1,44 +1,0 @@
-/*
- * Copyright 2018 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-apply plugin: 'java-library'
-
-jar {
-  baseName 'pantheon-queue'
-  manifest {
-    attributes(
-      'Specification-Title': baseName,
-      'Specification-Version': project.version,
-      'Implementation-Title': baseName,
-      'Implementation-Version': calculateVersion()
-      )
-  }
-}
-
-dependencies {
-  api project(':util')
-
-  compileOnly 'org.openjdk.jmh:jmh-generator-annprocess'
-
-  implementation project(':metrics')
-  implementation project(':services:util')
-
-  implementation 'org.apache.logging.log4j:log4j-api'
-  implementation 'com.google.guava:guava'
-  implementation 'org.rocksdb:rocksdbjni'
-
-  runtime 'org.apache.logging.log4j:log4j-core'
-
-  testImplementation 'junit:junit'
-  testImplementation 'org.assertj:assertj-core'
-}
