+++ settings.gradle
@@ -35,7 +35,7 @@ include 'ethereum:trie'
 include 'metrics'
 include 'pantheon'
 include 'services:kvstore'
-include 'services:queue'
+include 'services:tasks'
 include 'services:util'
 include 'testutil'
 include 'util'
